Separate states from component
