+++ react/components/TableCompose/index.js
@@ -1,6 +1,5 @@
-import React, { useEffect, useMemo, useReducer } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import reduce from 'lodash/reduce'
 import map from 'lodash/map'
 
 import Box from '../Box'
@@ -10,98 +9,14 @@ import FilterBar from '../FilterBar'
 
 import SimpleTable from './SimpleTable'
 import Toolbar from './Toolbar'
-import CheckboxContainer from './CheckboxContainer'
 import Totalizers from '../Totalizer'
 import BulkActions from './BulkActions'
 
-import {
-  getRowHeight,
-  getInitialHiddenFieldsFromSchema,
-  calculateTableHeight,
-  handleSelectionChange,
-} from './util'
+import { calculateTableHeight } from './util'
 
-const types = {
-  SET_DENSITY: 0,
-  DESELECT_ALL_LINES: 1,
-  SELECT_ALL_LINES: 2,
-  SELECT_LINE: 3,
-  HIDE_COLUMN: 4,
-  HIDE_ALL_COLUMNS: 5,
-  UNHIDE_ALL_COLUMNS: 6,
-}
+import useTableState from './state/useTableState'
 
-function reducer(state, action) {
-  switch (action.type) {
-    case types.SET_DENSITY: {
-      return {
-        ...state,
-        selectedDensity: action.density,
-        tableRowHeight: getRowHeight(action.density),
-      }
-    }
-    case types.DESELECT_ALL_LINES: {
-      return {
-        ...state,
-        selectedRows: [],
-        allLinesSelected: false,
-      }
-    }
-    case types.SELECT_ALL_LINES: {
-      return {
-        ...state,
-        selectedRows: action.selectedRows,
-        allLinesSelected: true,
-      }
-    }
-    case types.SELECT_LINE: {
-      return state.selectedRows.some(el => el.id === action.row.id)
-        ? {
-            ...state,
-            selectedRows: state.selectedRows.filter(
-              row => row.id !== action.row.id
-            ),
-            allLinesSelected: false,
-          }
-        : {
-            ...state,
-            selectedRows: [...state.selectedRows, action.row],
-          }
-    }
-    case types.HIDE_COLUMN: {
-      const index = state.hiddenFields.indexOf(action.key)
-
-      return index === -1
-        ? {
-            ...state,
-            hiddenFields: [...state.hiddenFields, action.key],
-          }
-        : {
-            ...state,
-            hiddenFields: state.hiddenFields.filter(
-              field => field !== action.key
-            ),
-          }
-    }
-    case types.HIDE_ALL_COLUMNS: {
-      return {
-        ...state,
-        hiddenFields: action.hiddenFields,
-      }
-    }
-    case types.UNHIDE_ALL_COLUMNS: {
-      return {
-        ...state,
-        hiddenFields: [],
-      }
-    }
-    default: {
-      return state
-    }
-  }
-}
-
-function Table({
+const Table = ({
   items,
   density,
   schema,
@@ -122,142 +37,22 @@ function Table({
   bulkActions,
   totalizers,
   filters,
-}) {
-  const [state, dispatch] = useReducer(reducer, {
-    tableRowHeight: getRowHeight(density),
-    selectedRows: [],
-    allLinesSelected: false,
-    selectedDensity: density,
-    hiddenFields: getInitialHiddenFieldsFromSchema(schema),
-  })
-
-  const hasPrimaryBulkAction =
-    bulkActions &&
-    bulkActions.main &&
-    typeof bulkActions.main.handleCallback === 'function'
-
-  const hasSecondaryBulkActions =
-    bulkActions.others && bulkActions.others.length > 0
-
-  const hasBulkActions = hasPrimaryBulkAction || hasSecondaryBulkActions
-
-  const data = useMemo(() => {
-    return hasBulkActions ? items.map((item, i) => ({ id: i, ...item })) : items
-  }, [items])
-
-  const staticSchema = useMemo(
-    () =>
-      hasBulkActions
-        ? {
-            ...schema,
-            properties: {
-              bulk: {
-                width: 40,
-                // eslint-disable-next-line
-                headerRenderer: () => {
-                  const selectedRowsLength = state.selectedRows.length
-                  const itemsLength = data.length
-
-                  const isChecked = selectedRowsLength === itemsLength
-                  const isPartial =
-                    selectedRowsLength > 0 && selectedRowsLength < itemsLength
-
-                  return (
-                    <CheckboxContainer
-                      checked={isChecked}
-                      onClick={handleSelectAllVisibleLines}
-                      id="all"
-                      partial={isPartial}
-                    />
-                  )
-                },
-                // eslint-disable-next-line
-                cellRenderer: ({ rowData }) => (
-                  <CheckboxContainer
-                    checked={state.selectedRows.some(
-                      row => row.id === rowData.id
-                    )}
-                    onClick={() => handleSelectLine(rowData)}
-                    id={rowData.id}
-                    disabled={state.allLinesSelected}
-                  />
-                ),
-              },
-              ...schema.properties,
-            },
-          }
-        : schema,
-    [state.selectedRows, state.allLinesSelected]
-  )
-
-  const displaySchema = useMemo(() => {
-    return {
-      ...staticSchema,
-      properties: reduce(
-        staticSchema.properties,
-        (acc, value, key) => {
-          if (state.hiddenFields.includes && state.hiddenFields.includes(key)) {
-            return acc
-          }
-          return { ...acc, [key]: value }
-        },
-        {}
-      ),
-    }
-  }, [state.hiddenFields, state.selectedRows])
-
-  useEffect(() => {
-    handleSelectionChange(
-      bulkActions,
-      state.allLinesSelected,
-      state.selectedRows
-    )
-  }, [state.selectedRows, state.allLinesSelected, bulkActions])
-
-  function handleToggleDensity(density) {
-    dispatch({ type: types.SET_DENSITY, density: density })
-  }
-
-  function toggleColumn(key) {
-    dispatch({ type: types.HIDE_COLUMN, key: key })
-  }
-
-  function onShowAllColumns() {
-    dispatch({ type: types.UNHIDE_ALL_COLUMNS })
-  }
-
-  function onHideAllColumns() {
-    dispatch({
-      type: types.HIDE_ALL_COLUMNS,
-      hiddenFields: Object.keys(staticSchema.properties),
-    })
-  }
-
-  function handleSelectAllLines() {
-    dispatch({ type: types.SELECT_ALL_LINES, selectedRows: data.selectedRows })
-  }
-
-  function handleSelectAllVisibleLines() {
-    if (
-      state.selectedRows.length <= data.length &&
-      state.selectedRows.length !== 0
-    ) {
-      handleDeselectAllLines()
-    } else {
-      dispatch({
-        type: types.SELECT_ALL_LINES,
-        selectedRows: data,
-      })
-    }
-  }
-
-  function handleSelectLine(rowData) {
-    dispatch({ type: types.SELECT_LINE, row: rowData })
-  }
-
-  function handleDeselectAllLines() {
-    dispatch({ type: types.DESELECT_ALL_LINES })
-  }
+}) => {
+  const {
+    state,
+    setDensity,
+    toggleColumn,
+    showAllColumns,
+    hideAllColumns,
+    selectAllRows,
+    deselectAllRows,
+    displaySchema,
+    data,
+    staticSchema,
+    hasBulkActions,
+    hasPrimaryBulkAction,
+    hasSecondaryBulkActions,
+  } = useTableState(schema, items, density, bulkActions)
 
   const properties = Object.keys(staticSchema.properties)
   const emptyState = !!(
@@ -269,11 +64,11 @@ function Table({
   const paginationClone = pagination ? Object.assign({}, pagination) : null
   if (paginationClone && hasBulkActions) {
     paginationClone.onNextClick = () => {
-      handleDeselectAllLines()
+      deselectAllRows()
       pagination.onNextClick()
     }
     paginationClone.onPrevClick = () => {
-      handleDeselectAllLines()
+      deselectAllRows()
       pagination.onPrevClick()
     }
   }
@@ -286,9 +81,11 @@ function Table({
           toolbar={toolbar}
           hiddenFields={state.hiddenFields}
           toggleColumn={toggleColumn}
-          handleHideAllColumns={onHideAllColumns}
-          handleShowAllColumns={onShowAllColumns}
-          handleToggleDensity={handleToggleDensity}
+          handleHideAllColumns={() =>
+            hideAllColumns(Object.keys(staticSchema.properties))
+          }
+          handleShowAllColumns={showAllColumns}
+          handleToggleDensity={setDensity}
           selectedDensity={state.selectedDensity}
           schema={staticSchema}
           actions={toolbar}
@@ -309,8 +106,8 @@ function Table({
         selectedRows={state.selectedRows}
         bulkActions={bulkActions}
         allLinesSelected={state.allLinesSelected}
-        onSelectAllLines={handleSelectAllLines}
-        onDeselectAllLines={handleDeselectAllLines}
+        onSelectAllLines={() => selectAllRows(data)}
+        onDeselectAllLines={deselectAllRows}
       />
 
       {emptyState ? (
