+++ react/components/TableCompose/state/actionTypes.js
@@ -1,0 +1,9 @@
+export default {
+  SET_DENSITY: 0,
+  DESELECT_ALL_LINES: 1,
+  SELECT_ALL_LINES: 2,
+  SELECT_LINE: 3,
+  HIDE_COLUMN: 4,
+  HIDE_ALL_COLUMNS: 5,
+  UNHIDE_ALL_COLUMNS: 6,
+}
